+++ core/src/main/java/com/cobber/fta/dates/DateTimeParserConfig.java
@@ -37,14 +37,14 @@ public class DateTimeParserConfig {
 	}
 
 	@Override
-	public boolean equals(Object obj) {
+	public boolean equals(final Object obj) {
 		if (this == obj)
 			return true;
 		if (obj == null)
 			return false;
 		if (getClass() != obj.getClass())
 			return false;
-		DateTimeParserConfig other = (DateTimeParserConfig) obj;
+		final DateTimeParserConfig other = (DateTimeParserConfig) obj;
 		return Objects.equals(locale, other.locale) && resolutionMode == other.resolutionMode
 				&& strictMode == other.strictMode;
 	}